Add dto for all
