+++ src/main/java/com/classhub/api/model/dto/Student/StudentDto.java
@@ -1,10 +1,10 @@
-package com.classhub.api.model.dto;
+package com.classhub.api.model.dto.Student;
 
 import lombok.Data;
 
 @Data
 public class StudentDto {
-
+    private Long id;
     private String username;
     private String role;
     private String first_name;
