+++ src/main/java/com/classhub/api/model/dto/StudentGrade/StudentGradeDto.java
@@ -1,0 +1,11 @@
+package com.classhub.api.model.dto.StudentGrade;
+
+import lombok.Data;
+
+@Data
+public class StudentGradeDto  {
+    private Long id;
+    private int grade;
+    private Long taskId;
+    private Long studentId;
+}
