+++ src/main/java/com/classhub/api/model/dto/Task/TaskCreationDto.java
@@ -1,0 +1,10 @@
+package com.classhub.api.model.dto.Task;
+
+import lombok.Data;
+
+@Data
+public class TaskCreationDto {
+    private String title;
+    private String description;
+    private Long teachingSubjectId;
+}
