+++ src/main/java/com/classhub/api/model/dto/TeacherSubject/TeacherSubjectDto.java
@@ -1,4 +1,4 @@
-package com.classhub.api.model.dto;
+package com.classhub.api.model.dto.TeacherSubject;
 
 import lombok.Data;
