+++ src/main/java/com/classhub/api/model/dto/TeachingSubject/TeachingSubjectCreationDto.java
@@ -1,0 +1,9 @@
+package com.classhub.api.model.dto.TeachingSubject;
+
+import lombok.Data;
+
+@Data
+public class TeachingSubjectCreationDto {
+    private Long subjectId;
+    private Long teachingPeriodId;
+}
