--- src/main/java/com/classhub/api/model/dto/TeachingSubjectDto.java
@@ -1,9 +1,0 @@
-package com.classhub.api.model.dto;
-
-import lombok.Data;
-
-@Data
-public class TeachingSubjectDto {
-    private Long subjectId;
-    private Long teachingPeriodId;
-}
